# Optimize encoding and decoding