--- a/go.mod
+++ b/go.mod
@@ -4,7 +4,6 @@ go 1.13
 
 require (
 	github.com/ipfs/go-cid v0.0.5 // indirect
-	github.com/multiformats/go-varint v0.0.5
 	github.com/stretchr/testify v1.4.0
 	github.com/whyrusleeping/cbor-gen v0.0.0-20200414195334-429a0b5e922e
 	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543