--- a/rle/bitvec.go
+++ b/rle/bitvec.go
@@ -34,14 +34,50 @@ var bitMasks = [9]byte{
 	0xFF,
 }
 
-func (bv *rbitvec) Get(count byte) byte {
-	res := byte(bv.bits) & bitMasks[count] // select count bits
-	bv.bits = bv.bits >> count             // remove those bits from storage
-	bv.bitCap = bv.bitCap - count          // decrease nuber of stored bits
+func (bv *rbitvec) GetByte() byte {
+	// Advancing byte by byte is simpler than advancing an odd number of
+	// bits because we _always_ load the next byte.
+	res := byte(bv.bits)
+	bv.bits >>= 8
 
 	if bv.index < len(bv.vec) { // if vector allows
 		// add bits onto the end of temporary storage
-		bv.bits = bv.bits | uint16(bv.vec[bv.index])<<bv.bitCap
+		bv.bits |= uint16(bv.vec[bv.index]) << (bv.bitCap - 8)
+	}
+
+	bv.index += 1
+	return res
+}
+
+func (bv *rbitvec) GetBit() bool {
+	// The specialized GetBit is easier for the compiler to optimize, for some reason.
+
+	res := (bv.bits&0x1 != 0)
+	bv.bits >>= 1
+	bv.bitCap -= 1
+
+	if bv.index < len(bv.vec) { // if vector allows
+		// add bits onto the end of temporary storage
+		bv.bits |= uint16(bv.vec[bv.index]) << bv.bitCap
+	}
+
+	// When we advance one by one, this branch is very predictable (and
+	// faster than fancy math).
+	if bv.bitCap < 8 {
+		bv.index += 1
+		bv.bitCap += 8
+	}
+	return res
+}
+
+func (bv *rbitvec) Get(count byte) byte {
+	res := byte(bv.bits) & bitMasks[count] // select count bits
+	bv.bits >>= count                      // remove those bits from storage
+	bv.bitCap -= count                     // decrease nuber of stored bits
+
+	if bv.index < len(bv.vec) { // if vector allows
+		// add bits onto the end of temporary storage
+		bv.bits |= uint16(bv.vec[bv.index]) << bv.bitCap
 	}
 
 	// Here be dragons
@@ -56,9 +92,9 @@ func (bv *rbitvec) Get(count byte) byte {
 
 	// if bitCap < 8 it underflows, then high bits get set to 1s
 	// we shift by 7 so the highest bit is in place of the lowest
-	inc := (bv.bitCap - 8) >> 7    // inc == 1 iff bitcap<8 (+10% perf)
-	bv.index = bv.index + int(inc) // increase index if we need more bits
-	bv.bitCap = bv.bitCap + inc*8  // increase bitCap by 8
+	inc := (bv.bitCap - 8) >> 7 // inc == 1 iff bitcap<8 (+10% perf)
+	bv.index += int(inc)        // increase index if we need more bits
+	bv.bitCap += inc * 8        // increase bitCap by 8
 
 	return res
 }