Overall, this gives us a ~18-20% speedup.

```
name            old time/op    new time/op    delta
DecodeEncode-4     115µs ± 1%      93µs ± 3%  -18.60%  (p=0.000 n=7+9)

name            old alloc/op   new alloc/op   delta
DecodeEncode-4    80.7kB ± 0%    80.7kB ± 0%     ~     (all equal)

name            old allocs/op  new allocs/op  delta
DecodeEncode-4      18.0 ± 0%      18.0 ± 0%     ~     (all equal)
```